+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# next
+
+* Update CSSO to [^4.0.0](https://github.com/css/csso/releases/tag/v4.0.0)
+* Plugin never produce a parse error, but drops bad parts of CSS instead
+* Drop support for Node.js < 8
+
 # 3.0.1
 
 * Drop dependency on deprecated `gulp-util` (#30, thanks to @TheDancingCode)